requirement.txt
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorlog==6.9.0
cron-descriptor==1.4.5
decorator==5.2.1
Django==4.2.23
django-celery-beat==2.8.1
django-filter==24.3
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.2.0
flower==2.0.1
humanize==4.10.0
idna==3.10
importlib_metadata==8.5.0
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.2
kombu==5.5.4
Markdown==3.7
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
prometheus_client==0.21.1
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.9.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.1.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tomli==2.2.1
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
zipp==3.20.2
